myproject/requirements.txt
Django==1.4.2
PIL==1.1.7
South==0.7.6
argparse==1.2.1
django-admin-bootstrapped==0.1
django-annoying==0.7.6
django-classy-tags==0.3.4.1
django-cms==2.3.4
django-debug-toolbar==0.9.4
django-extensions==1.0.1
django-filer==0.9.1
django-mptt==0.5.2
django-polymorphic==0.2
django-reversion==1.6.4
django-sekizai==0.6.1
easy-thumbnails==1.1
gunicorn==0.15.0
html5lib==0.95
psycopg2==2.4.5
wsgiref==0.1.2
